backend/requirements.txt
pipecat-ai[websocket,silero,deepgram]
anthropic
python-dotenv
loguru
fastapi
uvicorn[standard]
websockets